Renamed business logic file
